refactor: split gridView into multiple files
